Add file trait
